Sprint2 finished. Merging Main to Release.

+++ backend/client_app/views.py
@@ -5,6 +5,9 @@ from client_app.models import Client
 from client_app.serializers import ClientSerializer
 from django.shortcuts import get_object_or_404
 
+from company_app.models import Company
+from task_app.models import Task
+
 class ClientView(ApiView):
         
     def get(self, request, pk=None):
@@ -12,7 +15,7 @@ class ClientView(ApiView):
             client = get_object_or_404(Client, id=pk);
             serializer = ClientSerializer(client);
         else:
-            clients = Client.objects.all();
+            clients = Client.objects.filter(enabled=True);
             serializer = ClientSerializer(clients, many=True);
         return Response(serializer.data, status=status.HTTP_200_OK);
 
@@ -21,7 +24,7 @@ class ClientView(ApiView):
         if serializer.is_valid():
             serializer.save();
             return Response(serializer.data, status=status.HTTP_201_CREATED);
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST);
+        return Response({"message": "Hubo un error generando el cliente", "errors": serializer.errors}, status=status.HTTP_400_BAD_REQUEST);
 
     def put(self, request, pk=None):
         client = get_object_or_404(Client, id=pk);
@@ -29,15 +32,26 @@ class ClientView(ApiView):
         if serializer.is_valid():
             serializer.save();
             return Response(serializer.data);
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST);
+        return Response({"message": "Hubo un error modificando el cliente", "errors": serializer.errors}, status=status.HTTP_400_BAD_REQUEST);
 
     def delete(self, request, pk=None,):
         client = get_object_or_404(Client, id=pk);
         hard = request.query_params.get("hard", "false").lower() in ["true"]
         if hard:
+            if request.user.role != "admin":
+                raise self.permission_denied(request,
+                message="Solo admin puede realizar borrado físico."
+            );
             client.delete();
             return Response(status=status.HTTP_204_NO_CONTENT);
 
+        otherClients = Client.objects.filter(company=client.company);
+        if (otherClients.count() == 1):
+            Company.objects.filter(pk= client.company.id).update(enabled=False);
+        
+        relatedTask = Task.objects.filter(client=client, enabled=True);
+        relatedTask.update(enabled = False);
+        
         client.enabled = False
         client.save();
         return Response(status=status.HTTP_204_NO_CONTENT);
