+++ backend/company_app/views.py
@@ -2,8 +2,11 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 from django.shortcuts import get_object_or_404
+
+from client_app.models import Client
 from .models import Company
 from .serializers import CompanySerializer
+
 class CompanyView(APIView):
     
     def get(self, request, pk=None):
@@ -11,8 +14,7 @@ class CompanyView(APIView):
             company = get_object_or_404(Company, id=pk)
             serializer = CompanySerializer(company)
         else:
-            #companies = Company.objects.filter(enabled=True) para ver las eliminaciones sft
-            companies = Company.objects.all()
+            companies = Company.objects.filter(enabled=True)
             serializer = CompanySerializer(companies, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
@@ -22,7 +24,7 @@ class CompanyView(APIView):
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return Response({"message": "Hubo un error generando la empresa", "errors": serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
 
 
     def put(self, request, pk=None):
@@ -32,15 +34,25 @@ class CompanyView(APIView):
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return Response({"message": "Hubo un error modificando la empresa", "errors": serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
 
 
     def delete(self, request, pk=None):
         company = get_object_or_404(Company, id=pk)
         hard = request.query_params.get("hard", "false").lower() == "true"
         if hard:
+            if request.user.role != "admin":
+                raise self.permission_denied(request,
+                message="Solo admin puede realizar borrado físico."
+            );
             company.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)    
+        
+        relatedClients = Client.objects.filter(company=company, enabled=True);
+        
+        if (relatedClients.exists()):
+            return Response({"message": "No se puede eliminar una empresa que tiene clientes asociados. Elimine o reasigne los clientes primero."}, status=status.HTTP_400_BAD_REQUEST);
+        
         company.enabled = False
         company.save()
         return Response(status=status.HTTP_204_NO_CONTENT)
